Fixed FSPS problem
This is the problem where it showed a very blue spectrum with an old
age as a best fit to the stacks.

I think the issue was the FSPS library generating models where the
spectrum had values very close to zero ~10^-59. These are the spectra
where it has saved values that are very close to zero but not exactly
zero and that was throwing off the fitting code. Also these values
(which are bogus) have a blue slope look like a star-forming galaxy's
blue spectrum; almost like a blackbody.

As a two fold solution, while saving the fsps spectra to a fits file, I
had it skip over spectra where the median of the spectrum was less than
10^-30. I also made the age condition more stringent. Now, it allows
only spectra with ages less than or equal to 7.5 Gyr.